requirements.txt
# Core dependencies
openai>=1.0.0
ebooklib>=0.18.0
elevenlabs>=0.2.24
ffmpeg-python>=0.2.0
nltk>=3.8.1
mem0>=0.1.0  # Vector database client

# Utilities
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
tqdm>=4.66.1
colorama>=0.4.6

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0